requirements.txt
matlabengine
dill
numpy
scipy
python-dotenv
pydantic
PyAudio
SpeechRecognition[openai]
streamlit
langchain
langchain-openai
langchain-community
# langchain-anthropic
# langchain-mistralai
# langchain-cohere
# langchain-aws
# boto3
# langchain-google-genai